refactor: move Config type to value-objects and update imports
