+++ apps/io-session-manager-oi/src/domain/value-objects/config.vo.ts
